+++ src/main/java/com/acme/decorators/TotalMemoryDecorator.java
@@ -1,4 +1,4 @@
-package com.acme;
+package com.acme.decorators;
 
 public class TotalMemoryDecorator extends BaseServerStatus {
