+++ src/main/java/com/acme/facades/FacadeInterface.java
@@ -1,4 +1,4 @@
-package com.acme;
+package com.acme.facades;
 
 public interface FacadeInterface {
